♻️ Split date/time rendering into dumb components + a user-aware layer
The dumb components (src/components/datetime: Time, TimeRange, DateValue, RelativeTime) take locale/timeFormat/timeZone as props and format via the Intl layer in lib/localization — no context, usable server or client. UserTime/UserTimeRange (features/user) fill locale and the user's preferred timeFormat from the UserDTO, so only the user-aware wrapper touches the user context.

Retires LocalizationProvider/useLocalization: the (space) layout no longer mounts a provider, and scheduled-event-list renders via UserTime.
