+++ apps/web/src/app/[locale]/(space)/layout.tsx
@@ -4,10 +4,7 @@ import { TimeZoneChangeDetector } from "@/app/[locale]/timezone-change-detector"
 import { SessionRefresher } from "@/components/session-refresher";
 import { PayWall } from "@/features/billing/components/pay-wall";
 import { SpaceProvider } from "@/features/space/client";
-import { getLocale } from "@/i18n/server/get-locale";
 import { LocaleSync } from "@/lib/locale/client";
-import { LocalizationProvider } from "@/lib/localization/context";
-import { getLocaleDefaults } from "@/lib/localization/locales";
 import { getPathname } from "@/lib/pathname";
 import { createPrivateSSRHelper } from "@/trpc/server/create-ssr-helper";
 import { buildSafeRedirectUrl } from "@/utils/redirect";
@@ -19,10 +16,9 @@ export default async function Layout({
 }) {
   const helpers = await createPrivateSSRHelper();
 
-  const [user, space, locale] = await Promise.all([
+  const [user, space] = await Promise.all([
     helpers.user.getMe.fetch(),
     helpers.spaces.getCurrent.fetch(),
-    getLocale(),
   ]);
 
   if (!space) {
@@ -32,25 +28,15 @@ export default async function Layout({
     );
   }
 
-  const localeDefaults = getLocaleDefaults(locale);
-
   return (
     <HydrationBoundary state={dehydrate(helpers.queryClient)}>
       <SessionRefresher />
       <LocaleSync userLocale={user?.locale} />
       <TimeZoneChangeDetector initialTimeZone={user?.timeZone} />
-      <LocalizationProvider
-        defaults={{
-          locale,
-          timeFormat: user?.timeFormat ?? localeDefaults.timeFormat,
-          weekStart: user?.weekStart ?? localeDefaults.weekStart,
-        }}
-      >
-        <SpaceProvider>
-          {children}
-          <PayWall />
-        </SpaceProvider>
-      </LocalizationProvider>
+      <SpaceProvider>
+        {children}
+        <PayWall />
+      </SpaceProvider>
     </HydrationBoundary>
   );
 }
